+++ packages/@icon-magic/distribute/src/helpers/create-image-set.ts
