+++ packages/@icon-magic/distribute/src/helpers/create-sprite.ts
