+++ packages/@icon-magic/distribute/src/helpers/distribute-by-resolution.ts
