+++ packages/@icon-magic/distribute/src/helpers/distribute-svg.ts
