+++ packages/@icon-magic/distribute/src/helpers/utils.ts
